\name{AdvancedBSgenomeForge}

\alias{AdvancedBSgenomeForge}

\alias{forgeSeqlengthsRdsFile}
\alias{forgeSeqlengthsRdaFile}
\alias{forgeSeqFiles}
\alias{forgeMasksFiles}
\alias{forgeBSgenomeDataPkg}
\alias{forgeMaskedBSgenomeDataPkg}


\title{[Moved to BSgenomeForge] Tools to forge a BSgenome data package}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, the \code{forgeBSgenomeDataPkg}
  and \code{forgeMaskedBSgenomeDataPkg} functions are defined in the
  \pkg{BSgenomeForge} package.
}

\seealso{
  \code{BSgenomeForge::\link[BSgenomeForge]{forgeBSgenomeDataPkg}} in the
  \pkg{BSgenomeForge} package.
}

