\name{buildDataFiles}

\alias{buildDataFiles}

\title{A function to build the genome data files}

\description{
  This function is used to build the data files included
  in the "Biostring-based genome data packages" (aka "BSgenome
  packages") like the BSgenome.Hsapiens.UCSC.hg18 package.
  It's intended to be used by the maintainer of these
  packages, not by the end-user.
}

\usage{
  buildDataFiles(srcdir, destdir, names, prefix="", suffix="", comments=NULL, single.seq=TRUE)
}

\arguments{
  \item{srcdir}{
    [TODO: Document me]
  }
  \item{destdir}{
    [TODO: Document me]
  }
  \item{names}{
    [TODO: Document me]
  }
  \item{prefix}{
    [TODO: Document me]
  }
  \item{suffix}{
    [TODO: Document me]
  }
  \item{comments}{
    [TODO: Document me]
  }
  \item{single.seq}{
    [TODO: Document me]
  }
}

\details{
}

\value{
}

\author{H. Pages}

\seealso{}

\examples{
}

\keyword{classes}

