\name{buildDataFiles}

\alias{buildDataFiles}

\title{A function to build the genome data files}

\description{
  This function is used to build the data files included
  in the genome data packages like HsapiensGenome.hg18.
  It's intended to be used by the maintainer of these
  packages, not by the end-user.
}

\usage{
  buildDataFiles(srcdir, destdir, files, comments)
}

\arguments{
  \item{srcdir}{
    [TODO: Document me]
  }
  \item{destdir}{
    [TODO: Document me]
  }
  \item{files}{
    [TODO: Document me]
  }
  \item{comments}{
    [TODO: Document me]
  }
}

\details{
}

\value{
}

\author{H. Pages}

\seealso{}

\examples{
}

\keyword{classes}

