\name{getSeq}

\alias{getSeq}

\title{getSeq}

\description{
   Extract the sequence delimitted by a start and an end
   in a chromosome of a given genome.
}

\usage{
  getSeq(bsgenome, seqname, start=NA, end=NA)
}

\arguments{
  \item{bsgenome}{
  }
  \item{seqname}{
  }
  \item{start}{
  }
  \item{end}{
  }
}

\details{
}

\value{
  An object of the same type than \code{bsgenome[[seqname]]}.
}

\author{H. Pages}

\seealso{\link{BSgenome-class}, \link{seqnames}, \link[Biostrings]{subBString}}

\examples{
}

\keyword{manip}

